Rework loading of config.yaml
Move the config file loading to a new lib/pyhss_config.py file, instead
of having it duplicated in all files that load the config. This allows
replacing code like:

  import yaml

  try:
      with open("../config.yaml", 'r') as stream:
          config = (yaml.safe_load(stream))
  except:
      with open("config.yaml", 'r') as stream:
          config = (yaml.safe_load(stream))

with the following in all files:

  from pyhss_config import config

Replace code that loaded the config to self.config in various classes
with just using the global config variable. Adjust the few scripts that
had named it yaml_config instead of config to just config, so it is
consistent.

Let the new code not only load the config from the top dir of the git
repository, but also from a PYHSS_CONFIG env var, /etc/pyhss/config.yaml
and /usr/share/pyhss/config.yaml. This is useful for being able to still
load configs after installing PyHSS into a different location, for
example with pip or apt. It also allows running tests against PyHSS with
a config file in a separate directory (e.g. in osmo-ttcn3-hacks).