GSUP: SAI: send correct amount of auth vectors
Instead of having the amount of vectors hardcoded to one, implement the
same behavior as OsmoHLR: read num_vectors_req from the SAI request, and
return this amount or 5 if it is unset or higher than 5.

Add a "merge" parameter to GsupMessageBuilder.with_ie() and set it to
False for the auth_tuple IEs. This is needed, so multiple vectors are
actually in separate auth_tuple IEs, otherwise they would end up in the
same one.

With this patch, the following tests pass in the osmo-ttcn3-hacks HLR
testsuite:

* TC_gsup_sai
* TC_gsup_sai_num_auth_vectors