feat(docker): one container per service
PyHSS follows a microservice model and spreads it's usage over different
services. In a containerized world, it is considered good practice to
have one container per service instead of having a giant container that
does everthing.

While a monolithic container that does it all is convenient for an
enduser, a similar thing can be achived by leveraging the `compose`
feature of Docker and similar container runtimes.

Environment variables are substituted with a go envsubst variant. Using
that seemed like the easiest thing to do since pyHss itself does not
support env vars for config. Everything can be configured via
environment, but has defaults.