fix(docker): Ensure irrelevant dirs are omitted
This changes the `.dockerignore` file so that the directories themselves
are ignored. They were previously added to the image but were empty.
While it is not critical, this was not intentional.

The `docker` directory cannot be ignored completly because it contains
files needed during certain copy operations. That's why I delete it
afterwards from the file system. It is not ideal since the files will be
in the docker image layers. COPY --exclude=docker would be the solution
but that is not supported on the CI runners.

Also I've added the `.idea` directory, which is used by JetBrains IDEs
like PyCharm, which is what I use so I want it ignored.