fix(docker): Ensure log directory is app writable
The application still wants to create logfiles so we need to make sure
it is able to do so. The previous copy operation will create a log
directory but it will belong to root. I've also marked it as a VOLUME
because it contains variable data and added it to the compose file.

To do this, I had to move the USER directive to a later stage. To me,
putting it at the end seems to make the most sense.