chore(docker): improve IDE experience
This change makes it so that one can more easily configure docker
Python interpreters in IDEs like PyCharm.

* Rationale for moving the venv directory: You'll most likely want to
mount your source root to /opt/pyhss as a volume to use the container
runtime while also having up-to-date sources without rebuilding. But
that shadows the venv unless it is outside the source tree.

* Rationale for exec $@: IDEs will try to inject their own launch
CMDs to facilitate things like remote debugging the container. By
removing the CONTAINER_ROLE requirement, and defaulting to `exec $@`,
that can be achieved.