refactor(gsup): make SubscriberInfo generic + fix APNs
This makes loading the info for the subscriber that is used for ISD
requests generic. That will be useful later on as I want to implement
updating the subscriber info via the API, which will also trigger
ISD requests. That will make it easy to keep stick to the DRY principle.

As a drive-by: This also fixes APN info in the ISD. The previous
implementation would never have added any APN, which was clearly a bug.