tests: add pytest configs and fixtures
Prepare to run the tests automatically in CI. In order to do this, we
need (depending on the tests) to prepare a test database, run redis, and
run parts of PyHSS. Add pytest fixtures to accomplish this.

The following commits will adjust the test scripts to be able to run
with pytest in addition to directly running them with code in 'if
__name__ == '__main__':' blocks.

Pytest is compatible with unittest-based tests (used in the existing
API, milenage and diameter tests). For test_gsup_air, a test function
will be added that pytest will pick up.