tests/test_Diameter: rework test_B_Recv_AIR
Rework the test, so it does not fail with:
E       AssertionError: 16777251 != 0 : Application ID Mismatch

Before this patch, the AIR diameter packet is decoded, then the code
path in the if-condition is taken as can be seen by the log messages:

  DEBUG    UnitTestLogger:test_Diameter.py:54 Received request with Command Code: 318, ApplicationID: 16777251 and flags c0
  INFO     UnitTestLogger:test_Diameter.py:56 Received Request with command code 318 (3GPP Authentication-Information-Request) - Generating (AIA)
  INFO     UnitTestLogger:test_Diameter.py:59 Generated AIR

Part of the condition is packet_vars['ApplicationId'] == 16777251, which
is true. packet_vars does not get written back to, so the assert at the
end for packet_vars['ApplicationId'] == 0 cannot pass.

Change the test to just have one code path that receives the AIR packet
and ensures it gets decoded successfully. This matches the name of the
test (test_B_Recv_AIR).