feat(gsup): implement RAT restriction check
Uses the subscriber attributes to limit RAT types. You can restrict
the RAT by using the following subscriber attributes:

- rat_restriction_2g = forbidden
- rat_restriction_3g = forbidden
- rat_restriction_4g = forbidden

This feature will be implemented in diameter in a future commit as well,
allowing users to use PyHSS to restrict users to a certain RAT type.

Useful feature to have if you use PyHSS as a HLR for 2G / 3G and HSS for
4G on the same PLMN.