lib/database: AUC: add sqn_ind_bitlen
Prepare to implement the SQN counter as described in
3GPP TS 33.102 § C.1.1.2, instead of the currently used "sqn += 100"
approach. According to the spec, SQN should look like this:

  SQN = SEQ | IND

* SEQ is the number we increase with each new SQN.
* IND is an array index. The upcoming patches will assign one IND to
  each client connected to PyHSS (MSC/VLR, SGSN, MME, ...). This is the
  same logic as in OsmoHLR.

In order to do this, we need to know the subscriber-specific length of
IND. § C.3.2 recommends the value 5, but it can be different.

In the database, I use default=None instead of default=5, so we don't
need to add migration logic for existing subscriber entries. Upcoming
patches will check for None and use 5 in that case.