lib/{lte,milenage}: update copyright and license
Prepare to use SPDX-License-Identifiers across the whole source tree.
These two files are special, because they have the following copyright
note:

  Copyright (c) 2016-present, Facebook, Inc.
  All rights reserved.

  This source code is licensed under the BSD-style license found in the
  LICENSE file in the root directory of this source tree. An additional grant
  of patent rights can be found in the PATENTS file in the same directory.

This is a non-standard text as it says copyright "-present" and mentions
patents. However Facebook has imported the original files into the Magma
project [1] and changed the copyright to the following:

  Copyright 2020 The Magma Authors.

  This source code is licensed under the BSD-style license found in the
  LICENSE file in the root directory of this source tree.

  Unless required by applicable law or agreed to in writing, software
  distributed under the License is distributed on an "AS IS" BASIS,
  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  See the License for the specific language governing permissions and
  limitations under the License.

The LICENSE file was changed from a custom BSD license to a BSD 3-Clause
License [2]. Use this newer copyright and license, so we can just refer
to the license as "SPDX-License-Identifier: BSD-3-Clause", which has the
same text [3].

The lib/milenage.py file has been modified by additional authors, add
their copyright to the file as well.

[1]: https://github.com/magma/magma/blob/509e5b4bcdb68b0021044fb1dc8cf47554446b8f/lte/gateway/python/magma/subscriberdb/crypto/milenage.py#L1-L12
[2]: https://github.com/magma/magma/commit/47dfed175ee0615c020ef32e40c39276710a0036
[3]: https://spdx.org/licenses/BSD-3-Clause.html