config: database: default to sqlite
The SQLite database is a more suitable default, as it works fine for
small scale and development scenarios while it does not need additional
installation and configuration of database software.

For big deployments, users will need to adjust their config anyway. With
new commented out blocks for mysql and postgresql, switching to those is
trivial.

Existing users are not affected by this change, as they practically need
to have their own version of the config with their own database
information already.