DB: let only the main service create/upgrade it
Before this patch we have multiple services trying to create the
database and trying to upgrade the schema at the same time. (The logic
for upgrading is currently just creating missing tables, I plan to
improve this with future patches.)

Instead of doing that, let only the main service, pyhss_hss, do this and
let all other services wait until the database schema is ready.