DB: add database_schema_version table
Prepare to store the current DB version inside the database itself, and
to make automatic database upgrades possible. With this, users will be
able to just use the latest version without worrying about manually
adjusting the database.

There are currently scripts for updating the database, but they are
fairly limited:

> Database upgrades are currently limited to semi-automation.
>
> Alembic is used to handle database schema upgades.
>
> This will not give a foolproof upgrade, ensure you read the generated
> scripts. For best results (and in production environments), read
> lib/database.py and compare each base object to the table in your
> database. Types for columns should also be checked.