DB: refuse start with schema below release 1.0.1
Adding automatic database migrations from 1.0.0 to 1.0.1 is not feasible
because the type of ims_subscriber.ifc_path was changed from VARCHAR(18)
to VARCHAR(512), and this is not trivial to do with database type sqlite
(there is no ALTER TABLE MODIFY). I don't think many people are running
1.0.0 from 2023, if any at all. So just document the changes that were
made, check if the user's database is from before 1.0.1 and if so, point
at this documentation.