DB: rework logic for initializing an empty db
Do not unconditionally create missing tables with the current scheme
when opening the database anymore. This is a potential race condition
when multiple services open the database.

Instead of doing that, detect when the database is completely empty and
only create all tables with the current scheme in that case. A database
is typically empty when using MySQL/PostgreSQL and creating DBs with a
root account that has more privileges, before allowing applications to
access the DB with a separate, less privileged account.