feat: implement inbound roaming reject
Implements inbound roaming reject cause configuration. This is useful
for private networks and public SIMs from other carriers try to join.

In that case, you want to respond with ROAMING_NOT_ALLOWED as reject
cause in 2G/3G/4G. Otherwise, the UE may believe that its subscription
is terminated.

Defaults to ROAMING_NOT_ALLOWED for a safe OOBE. Production carriers are
advised to change this to IMSI_UNKNOWN.